change type of TestScore from int to float

+++ API/Dtos/Lesson/AddLessonDto.cs
@@ -13,7 +13,7 @@ namespace API.Dtos.Lesson
         public string UrlPractice { get; set; }
         public int Number { get; set; }
         public int Importance { get; set; } = 0;
-        public int TestScore { get; set; } = -1;
+        public float TestScore { get; set; } = -1;
         public bool IsTheoryCompleted { get; set; } = false;
         public bool IsPracticeCompleted { get; set; } = false;
     }
